feat: add more information when reading directory
`fs.readdirSync(p, { withFileTypes: true })` - return the Dirent object, that conveniently has isDirectory, isFile, isSymbolic link methods, which in return allow us not to do additional lstat when deciding which class to create.

In some cases, File class do require stats to be calculated, which will happen as soon as needed and will be cached.